Codons work correctly.
It seems that the question is a little more in depth than I expected.
All protein strings need to be generated even if they are contained
inside another one. I did not account for that. It also means that
some strings would be repetitive, so putting the answers into a hash
would be a good idea.